Fix startup hang: OpenJDK No monotonic clock was available + 100% CPU (#4)
The latest Ubuntu image cause container to hang on startup, with 100% CPU, with error:
"OpenJDK Client VM warning: No monotonic clock was available"

I've investigated in #4 and it appears to be related to libc (part of the base Ubuntu image).

For now, to get 3.25.1 out (contains important security fix), I've chosen to lock it to a known good base image version.

It's the lesser of 2 evils.

Hopefully, by the next release, Ubuntu issue will be fixed.